support crdb

+++ src/entity/commodity.rs
@@ -7,9 +7,9 @@ use sea_orm::{entity::prelude::*, ActiveValue::NotSet, Set};
 #[sea_orm(table_name = "tiny_orders_commodity")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = true)]
-    pub id: u32,
+    pub id: i64,
     pub title: String,
-    pub price: u32,
+    pub price: i64,
     pub description: String,
     pub updated_at: DateTime,
     pub created_at: DateTime,
