+++ src/entity/evaluation.rs
@@ -6,9 +6,9 @@ use sea_orm::{entity::prelude::*, ActiveValue::NotSet, Set};
 #[sea_orm(table_name = "tiny_orders_evaluation")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = true)]
-    pub id: u32,
-    pub consumer_id: u32,
-    pub commodity_id: u32,
+    pub id: i64,
+    pub consumer_id: i64,
+    pub commodity_id: i64,
     pub evaluation: String,
     pub updated_at: DateTime,
     pub created_at: DateTime,
@@ -20,7 +20,7 @@ pub enum Relation {}
 impl ActiveModelBehavior for ActiveModel {}
 
 impl ActiveModel {
-    pub fn rand_fake_new(consumer_id: u32, commodity_id: u32) -> Self {
+    pub fn rand_fake_new(consumer_id: i64, commodity_id: i64) -> Self {
         let create_at = Local::now().naive_local();
         Self {
             consumer_id: Set(consumer_id),
